+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+- HomeView UI/UX for courses / lessons
+
 ## [1.2.14] - 2022-04-12
 
 ### Added